requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
transformers>=4.41,<5
accelerate>=0.30
sentencepiece>=0.1.99
